fixed basic asset loading, next is fixing skyboxes

+++ Engine/Source/Systems/ECS/Components/MeshComponent.h
@@ -9,8 +9,8 @@ namespace Aozora {
 	struct MeshComponent {
 	public:
 		bool m_isActive{ true };
-		uint32_t meshID{0};
-		uint32_t materialID{ 0 };
+		uint64_t meshID{0};
+		uint64_t materialID{ 0 };
 
 		template<class Archive>
 		void serialize(Archive& archive) {
